chore: move help drawer to DashboardLayout
